refactor: extract components gallery state
